hopefully finished database init

+++ create_directories.sh
@@ -11,16 +11,31 @@ echo -e "#******************************************************************"
 
 read -p 'Project name - this is used as a directory name, so must be conformant to bash requirements : ' PROJECT_NAME
 read -p 'Path to code (the django_artisan folder where manage.py resides) : ' CODE_PATH
-read -p 'Absolute path to User home dir : ' USER_DIR
+read -p "Absolute path to User home dir : " USER_DIR
 read -p 'User account name : ' USER
 
 mkdir -p /etc/opt/${PROJECT_NAME}/settings
 mkdir -p /etc/opt/${PROJECT_NAME}/static_files
 mkdir -p ${USER_DIR}/${PROJECT_NAME}/logs
 
-ln -s ${CODE_PATH} /opt/${PROJECT_NAME}
-ln -s ${CODE_PATH}/static/ /etc/opt/${PROJECT_NAME}/static_files/static
-ln -s ${CODE_PATH}/media/ /etc/opt/${PROJECT_NAME}/static_files/media
+if [[ -L /opt/${PROJECT_NAME} ]]
+then
+    echo "**WARNING** /opt/${PROJECT_NAME} exists already!"
+else
+    ln -s ${CODE_PATH} /opt/${PROJECT_NAME}
+fi
+if [[ -L /etc/opt/${PROJECT_NAME}/static_files/static ]]
+then
+    echo "**WARNING** /etc/opt/${PROJECT_NAME}/static_files/static exists!"
+else
+    ln -s ${CODE_PATH}/static/ /etc/opt/${PROJECT_NAME}/static_files/static
+fi
+if [[ -L /etc/opt/${PROJECT_NAME}/static_files/media ]]
+then
+    echo "**WARNING** /etc/opt/${PROJECT_NAME}/static_files/media exists!"
+else
+    ln -s ${CODE_PATH}/media/ /etc/opt/${PROJECT_NAME}/static_files/media
+fi
 
 sudo chcon -R -t container_file_t /opt/${PROJECT_NAME}
 
@@ -38,5 +53,6 @@ then
 fi
 
 SCRIPTPATH="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
-echo $PROJECT_NAME > ${SCRIPTPATH}/.proj
+echo "PROJECT_NAME=${PROJECT_NAME}" > ${SCRIPTPATH}/.proj
+echo "CODE_PATH=${CODE_PATH}" >> ${SCRIPTPATH}/.proj
 chown ${USER}:${USER} ${SCRIPTPATH}/.proj
